+++ NEWS.md
@@ -1,7 +1,8 @@
-# lest 1.0.1
+# lest 1.1.0
 
   * Added `exceeds_tumbling_sum()` for checking when a partial cummulative sum
-  ("tumbling sum"") hits a certain threshold value. Thanks to
+  ("tumbling sum"") hits a certain threshold value. This is for example useful
+  for filtering geographic positions by a minimum distance. Thanks to
   [@klmr](https://github.com/klmr) for helping me come up with a name.
     
   * Added plain C clones of `dplyr::cumall()` and `dplyr::cumany()`